Tests for alias behavior with FiledNameConverter
Closes #275
